feat dockerfile build & run + add exeption

+++ scripts/after-deploy.sh
@@ -4,12 +4,17 @@ echo "> USER-APP REPOSITORY: $REPOSITORY"
 
 cd $REPOSITORY
 
-# 기존 컨테이너 중지 및 삭제
-docker stop user-app
-docker rm user-app
+# 기존 컨테이너 확인 후 중지 및 삭제
+if docker ps -a | grep -q user-app; then
+    docker stop user-app
+    docker rm user-app
+fi
 
 # Docker Build
-docker build -t user-app:latest .
-
-# Docker Run
-docker run -d -p 8000:8000 --name user-app user-app:latest
+if docker build -t user-app:latest .; then
+    # Docker Run
+    docker run -d -p 8000:8000 --name user-app user-app:latest
+else
+    echo "Docker build failed."
+    exit 1
+fi
